refactor: melhorando a estrutura dos folders
