--- Cliente/.gitignore
@@ -1,12 +1,0 @@
-/Cliente.class
-/ClienteCreateFrame$1.class
-/ClienteCreateFrame$ButtonClickListener.class
-/ClienteCreateFrame.class
-/ClienteReadFrame$1.class
-/ClienteReadFrame.class
-/ClienteUpdateFrame.class
-/ClienteUpdateFrame$1.class
-/ClienteUpdateFrame$ButtonClickListener.class
-/ClienteDeleteFrame.class
-/ClienteDeleteFrame$ButtonClickListener.class
-/ClienteDeleteFrame$1.class
