--- Pedido/.gitignore
@@ -1,4 +1,0 @@
-/Pedido.class
-/PedidoCreateFrame.class
-/PedidoCreateFrame$1.class
-/PedidoCreateFrame$ButtonClickListener.class
