--- Pizza/.gitignore
@@ -1,1 +1,0 @@
-/Pizza.class
