+++ Telas/Cliente/.gitignore
@@ -1,0 +1,8 @@
+/ClienteCreateFrame.class
+/ClienteDeleteFrame$1.class
+/ClienteDeleteFrame.class
+/ClienteReadFrame$1.class
+/ClienteReadFrame.class
+/ClienteUpdateFrame.class
+/ClienteCreateFrame$1.class
+/ClienteUpdateFrame$1.class
