+++ Telas/FormaDePagamento/.gitignore
@@ -1,10 +1,8 @@
-/FormaDePagamentoRead.class
-/FormaDePagamentoReadFrame.class
-/FormaDePagamentoCreateFrame.class
-/FormaDePagamento.class
 /FormaDePagamentoCreateFrame$1.class
 /FormaDePagamentoCreateFrame$ButtonClickListener.class
+/FormaDePagamentoCreateFrame.class
 /FormaDePagamentoReadFrame$1.class
-/FormaDePagamentoUpdateFrame.class
+/FormaDePagamentoReadFrame.class
 /FormaDePagamentoUpdateFrame$1.class
 /FormaDePagamentoUpdateFrame$ButtonClickListener.class
+/FormaDePagamentoUpdateFrame.class
