+++ Telas/Pedido/.gitignore
@@ -1,0 +1,2 @@
+/PedidoCreateFrame$1.class
+/PedidoCreateFrame.class
